Move clientapi/core scripts to server-side - modified to include prepared header/footer for each
