+++ core/resources/scripts/labkey/Ajax.js
@@ -1,21 +1,11 @@
-/**
- * @fileOverview
- * @author <a href="https://www.labkey.org">LabKey</a> (<a href="mailto:info@labkey.com">info@labkey.com</a>)
- * @license Copyright (c) 2012-2019 LabKey Corporation
- * <p/>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p/>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p/>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * <p/>
+/*
+ * Copyright (c) 2011-2020 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
+
+var LABKEY = require("./init");
+LABKEY.ExtAdapter = require("ExtAdapter").Adapter;
 
 /**
  * @namespace
@@ -259,5 +249,16 @@ LABKEY.Ajax = new function () {
 
             return xhr;
         }
+    }
+};
+
+// Rhino does not support XMLHttpRequest, so here we wrap the one from our ExtAdapter.
+exports.Ajax = {
+    request : function (config)
+    {
+        var o = LABKEY.ExtAdapter.Ajax.request(config);
+        if (LABKEY.ExtAdapter.isObject(o))
+            return o.responseJSON || o.responseXML || o.responseText;
+        return o;
     }
 };
