+++ core/resources/scripts/labkey/Query.js
@@ -1,21 +1,15 @@
-/**
- * @fileOverview
- * @author <a href="https://www.labkey.org">LabKey</a> (<a href="mailto:info@labkey.com">info@labkey.com</a>)
- * @license Copyright (c) 2012-2019 LabKey Corporation
- * <p/>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p/>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p/>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * <p/>
+/*
+ * Copyright (c) 2011-2020 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
+var console = require("console");
+var LABKEY = require("./init");
+LABKEY.ActionURL = require("./ActionURL").ActionURL;
+LABKEY.Ajax = require("./Ajax").Ajax;
+LABKEY.ExtAdapter = require("ExtAdapter").Adapter;
+LABKEY.Filter = require("./Filter").Filter;
+LABKEY.Utils = require("./Utils").Utils;
 
 /**
  * @namespace Query static class to programmatically retrieve, insert, update and
@@ -2423,927 +2417,4 @@ LABKEY.Query.Filter.DoesNotHaveMissingValue.prototype = new LABKEY.Query.Filter;
     };
 })();
 
-/**
-* @name LABKEY.Query.ModifyRowsOptions
-* @class   ModifyRowsOptions class to describe
-            the third object passed to the successCallback function
-            by {@link LABKEY.Query.updateRows}, {@link LABKEY.Query.insertRows} or
-            {@link LABKEY.Query.deleteRows}. This object's properties are useful for
-            matching requests to responses, as HTTP requests are typically
-            processed asynchronously.
- *            <p>Additional Documentation:
- *              <ul>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
- *                      How To Find schemaName, queryName &amp; viewName</a></li>
- *              </ul>
- *           </p>
-  * @see LABKEY.Query.updateRows
-  * @see LABKEY.Query.insertRows
-  * @see LABKEY.Query.deleteRows
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.ModifyRowsOptions#
-* @field
-*/
-
-/**
-* @name headers
-* @description  An object containing one property for each HTTP header sent to the server.
-* @type Object
-*/
-
-/**
-* @name method
-* @description The HTTP method used for the request (typically 'GET' or 'POST').
-* @type String
- */
-
-/**
-* @name url
-* @description  The URL that was requested.
-* @type String
-*/
-
-/**
-* @name jsonData
-* @description  The data object sent to the server. This will contain the following properties:
-          <ul>
-            <li><b>schemaName</b>: String. The schema name being modified.  This is the same schemaName
-                the client passed to the calling function. </li>
-            <li><b>queryName</b>: String. The query name being modified. This is the same queryName
-                the client passed to the calling function.  </li>
-            <li><b>rows</b>: Object[]. Array of row objects that map the names of the row fields to their values.
-            The fields required for inclusion for each row depend on the which LABKEY.Query method you are
-            using (updateRows, insertRows or deleteRows). </li>
-         </ul>
- <p/>
- <b>For {@link LABKEY.Query.updateRows}:</b> <p/>
- For the 'updateRows' method, each row in the rows array must include its primary key value
- as one of its fields.
- <p/>
- An example of a ModifyRowsOptions object for the 'updateRows' successCallback:
- <pre name="code" class="xml">
- {"schemaName": "lists",
-  "queryName": "API Test List",
-  "rows": [
- {"Key": 1,
- "FirstName": "Z",
- "Age": "100"}]
- } </pre></code>
-
- <p/>
- <b>For {@link LABKEY.Query.insertRows}:</b> <p/>
- For the 'insertRows' method, the fields of the rows should look the same as
- they do for the 'updateRows' method, except that primary key values for new rows
- need not be supplied if the primary key columns are auto-increment.
- <p/>
- An example of a ModifyRowsOptions object for the 'insertRows' successCallback:
- <pre name="code" class="xml">
- {"schemaName": "lists",
-  "queryName": "API Test List",
-  "rows": [
-  {"FirstName": "C",
- "Age": "30"}]
- } </pre></code>
-
- <p/>
- <b>For {@link LABKEY.Query.deleteRows}:</b> <p/>
- For the 'deleteRows' method, the fields of the rows should look the
- same as they do for the 'updateRows' method, except that the 'deleteRows'
- method needs to supply only the primary key values for the rows. All
- other row data will be ignored.
- <p/>
- An example of a ModifyRowsOptions object for the 'deleteRows' successCallback:
- <pre name="code" class="xml">
- {"schemaName": "lists",
-  "queryName": "API Test List",
-  "rows": [
-{"Key": 3}]
- } </pre></code>
-* @type  Object
-*/
-
-/**#@-*/
-
- /**
-* @name LABKEY.Query.ModifyRowsResults
-* @class  ModifyRowsResults class to describe
-            the first object passed to the successCallback function
-            by {@link LABKEY.Query.updateRows}, {@link LABKEY.Query.insertRows} or
-            {@link LABKEY.Query.deleteRows}. This object's properties are useful for
-            matching requests to responses, as HTTP requests are typically
-            processed asynchronously.
-  *            <p>Additional Documentation:
-  *              <ul>
-  *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
-  *                      How To Find schemaName, queryName &amp; viewName</a></li>
-  *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=javascriptTutorial">LabKey JavaScript API Tutorial</a> and
-  *                      <a href="https://www.labkey.org/home/Study/demo/wiki-page.view?name=reagentRequest">Demo</a></li>
-  *              </ul>
-  *           </p>
-* @example For example:
-        <pre name="code" class="xml">
-        {  "schemaName": "lists",
-           "queryName": "API Test List"
-           "rowsAffected": 1,
-           "command": "insert",
-           "errors": [],
-           "rows": [{ Key: 3, StringField: 'NewValue'}]
-        } </pre></code>
-* @see LABKEY.Query.updateRows
-* @see LABKEY.Query.insertRows
-* @see LABKEY.Query.deleteRows
-* @see LABKEY.Query.saveRows
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.ModifyRowsResults#
-* @field
-*/
-
-/**
-* @name LABKEY.Query.ModifyRowsResults#schemaName
-* @description  Contains the same schemaName the client passed to the calling function.
-* @type  String
-*/
-
-/**
-* @name     LABKEY.Query.ModifyRowsResults#queryName
-* @description  Contains the same queryName the client passed to the calling function.
-* @type     String
-*/
-
-/**
-* @name    command
-* @description   Will be "update", "insert", or "delete" depending on the API called.
-* @type    String
-*/
-
-/**
-* @name    errors
-* @description   Objects will contain the properties 'id' (the field to which the error is related, if any),
-*                and 'msg' (the error message itself).
-* @type    Array
-*/
-
-/**
-* @name  rowsAffected
-* @description  Indicates the number of rows affected by the API action.
-            This will typically be the same number of rows passed in to the calling function.
-* @type        Integer
-*/
-
-/**
-* @name LABKEY.Query.ModifyRowsResults#rows
-* @description   Array of rows with field values for the rows updated, inserted,
-            or deleted, in the same order as the rows supplied in the request. For insert, the
-            new key value for an auto-increment key will be in the returned row's field values.
-            For insert or update, the other field values may also be different than those supplied
-            as a result of database default expressions, triggers, or LabKey's automatic tracking
-            feature, which automatically adjusts columns of certain names (e.g., Created, CreatedBy,
-            Modified, ModifiedBy, etc.).
-* @type    Object[]
-*/
-
-/**#@-*/
-
-/**
-* @name LABKEY.Query.SelectRowsOptions
-* @class  SelectRowsOptions class to describe
-           the third object passed to the successCallback function by
-           {@link LABKEY.Query.selectRows}.  This object's properties are useful for
-           matching requests to responses, as HTTP requests are typically
-           processed asynchronously.
- *            <p>Additional Documentation:
- *              <ul>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
- *                      How To Find schemaName, queryName &amp; viewName</a></li>
- *              </ul>
- *           </p>
-* @see LABKEY.Query.selectRows
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.SelectRowsOptions#
-* @field
-*/
-
-/**
-* @name   query.schemaName
-* @description   Contains the same schemaName the client passed to the
-            calling function. See <a class="link"
-            href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
-            How To Find schemaName, queryName &amp; viewName</a>.
-* @type   String
-*/
-
-/**
-* @name   query.queryName
-* @description   Contains the same queryName the client passed
-            to the calling function. See
-            <a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
-            How To Find schemaName, queryName &amp; viewName</a>.
-* @type   String
-*/
-
-/**
-* @name    query.viewName
-* @description 	Name of a valid custom view for the chosen queryName. See
-            <a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
-            How To Find schemaName, queryName &amp; viewName</a>.
-* @type    String
-*/
-
-/**
-* @name    query.offset
-* @description  Row number at which results should begin.
-            Use this with maxRows to get pages of results.
-* @type   Integer
-*/
-
-/**
-* @name   query.sort
-* @description	Sort specification. This can be a comma-delimited list of
-            column names, where each column may have an optional dash (-) before the name
-            to indicate a descending sort.
-* @type  String
-*/
-
-/**
-* @name   maxRows
-* @description   Maximum number of rows to return
-* @type     Integer
-*/
-
-/**
-* @name   filters
-* @description   An object whose properties are filter specifications, one for each filter you supplied.
- *          All filters are combined using AND logic. Each one is of type {@link LABKEY.Filter.FilterDefinition}.
- *          The list of valid operators:
-            <ul><li><b>eq</b> = equals</li>
-            <li><b>neq</b> = not equals</li>
-            <li><b>gt</b> = greater-than</li>
-            <li><b>gte</b> = greater-than or equal-to</li>
-            <li><b>lt</b> = less-than</li>
-            <li><b>lte</b> = less-than or equal-to</li>
-            <li><b>dateeq</b> = date equal</li>
-            <li><b>dateneq</b> = date not equal</li>
-            <li><b>neqornull</b> = not equal or null</li>
-            <li><b>isblank</b> = is null</li>
-            <li><b>isnonblank</b> = is not null</li>
-            <li><b>contains</b> = contains</li>
-            <li><b>doesnotcontain</b> = does not contain</li>
-            <li><b>startswith</b> = starts with</li>
-            <li><b>doesnotstartwith</b> = does not start with</li>
-            <li><b>in</b> = equals one of</li></ul>
-* @type    Object
-*/
-
-/**#@-*/
-
-/**
-* @name LABKEY.Query.FieldMetaDataLookup
-* @class  Lookup metadata about a single field.
- *            <p>Additional Documentation:
- *              <ul>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
- *                      How To Find schemaName, queryName &amp; viewName</a></li>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=javascriptTutorial">LabKey JavaScript API Tutorial</a> and
- *                      <a href="https://www.labkey.org/home/Study/demo/wiki-page.view?name=reagentRequest">Demo</a></li>
- *              </ul>
- *           </p>
- * @see LABKEY.Query.FieldMetaData
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name    containerPath
-* @description The path to the container that this lookup points to, if not the current container
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name    public
-* @description Whether the target of this lookup is exposed as a top-level query
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name queryName
-* @description The name of the query that this lookup targets
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name schemaName
-* @description The name of the schema that this lookup targets
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name keyColumn
-* @description The name of column in the lookup's target that will be joined to the value in the local field
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaDataLookup#
-* @field
-* @name displayColumn
-* @description The name of the column in the lookup's target that will be shown as its value, instead of the raw key value
-* @type    String
-*/
-
-/**#@-*/
-
-/**
-* @name LABKEY.Query.FieldMetaData
-* @class  Metadata about a single field.
- *            <p>Additional Documentation:
- *              <ul>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
- *                      How To Find schemaName, queryName &amp; viewName</a></li>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=javascriptTutorial">LabKey JavaScript API Tutorial</a> and
- *                      <a href="https://www.labkey.org/home/Study/demo/wiki-page.view?name=reagentRequest">Demo</a></li>
- *              </ul>
- *           </p>
- * @see LABKEY.Query.selectRows
- * @see LABKEY.Query.getQueryDetails
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    name
-* @description The name of the field
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    friendlyType
-* @description A friendlier, more verbose description of the type, like "Text (String)" or "Date and Time"
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    shownInInsertView
-* @description Whether this field is intended to be displayed in insert UIs
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    shownInDetailView
-* @description Whether this field is intended to be displayed in detail UIs
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    shownInUpdateView
-* @description Whether this field is intended to be displayed in update UIs
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name    versionField
-* @description Whether this field's value stores version information for the row
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name userEditable
-* @description Whether this field is intended to be edited directly by the user, or managed by the system
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name calculated
-* @description Whether this field is a calculated value such as something generated by a SQL expression,
-* or a "real"/"physical" column in the database
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name readOnly
-* @description Whether the field's value can be modified
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name nullable
-* @description Whether the field's value is allowed to be null
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name mvEnabled
-* @description Whether this field supports missing value indicators instead of or addition to its standard value
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name keyField
-* @description Whether this field is part of the row's primary key
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name hidden
-* @description Whether this value is intended to be hidden from the user, especially for grid views
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name autoIncrement
-* @description Whether this field's value is automatically assigned by the server, like a RowId whose value is determined by a database sequence
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name jsonType
-* @description The type of JSON object that will represent this field's value: string, boolean, date, int, or float
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name importAliases
-* @description Alternate names for this field that may appear in data when importing, whose values should be mapped to this field
-* @type    String[]
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name tsvFormat
-* @description The format string to be used for TSV exports
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name format
-* @description The format string to be used for generating HTML
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name excelFormat
-* @description The format string to be used for Excel exports
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name extFormat
-* @description The format string that can be passed to Ext components.  This is currently only supported for dates.
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name extFormatFn
-* @description A function that can be used to produce the formatted string for this field.  This is currently supported for dates and numeric values.
-* Note: this function is returned as a string, so you will need to evaluate it to convert it to a function.  See example below.
-* @example <pre name="code" class="xml">
-* var formatFn = eval(meta.extFormatFn);
-* var formattedValue = formatFn(data);
-* </pre></code>
-*
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name caption
-* @description The caption to be shown for this field, typically in a column header, which may differ from its name
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name shortCaption
-* @description The caption for this field, without any prefix from potential parent lookups. In many cases this will be identical to the caption property.
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name description
-* @description The description for this field
-* @type    String
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name inputType
-* @description The type of form input to be used when editing this field, such as select, text, textarea, checkbox, or file
-* @type    boolean
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.FieldMetaData#
-* @field
-* @name lookup
-* @description Information about this field's lookup configuration
-* @type    LABKEY.Query.FieldMetaDataLookup
-*/
-
-/**#@-*/
-
-/**
-* @name LABKEY.Query.SelectRowsResults
-* @class  SelectRowsResults class to describe the first
-            object passed to the successCallback function by
-            {@link LABKEY.Query.selectRows}. This object's properties are useful for
-            matching requests to responses, as HTTP requests are typically
-            processed asynchronously.
- *            <p>Additional Documentation:
- *              <ul>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
- *                      How To Find schemaName, queryName &amp; viewName</a></li>
- *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=javascriptTutorial">LabKey JavaScript API Tutorial</a> and
- *                      <a href="https://www.labkey.org/home/Study/demo/wiki-page.view?name=reagentRequest">Demo</a></li>
- *              </ul>
- *           </p>
-* @see LABKEY.Query.selectRows
-*/
-
-/**#@+
-* @memberOf LABKEY.Query.SelectRowsResults#
-* @field
-*/
-
-/**
-* @name    schemaName
-* @description the name of the resultset's source schema.
-* @type    String
-*/
-
-/**
-* @name    queryName
-* @description the name of the resultset's source query.  In some cases, such as an 'executeSql' call with 'saveInSession' set to true, the
- * query name may refer to temporary query that can be used to re-retrieve data for the duration of the user's session. 
-* @type    String
-*/
-
-/**
-* @name    metaData
-* @description Contains type and lookup information about the columns in the resultset.
-* @type    Object
-*/
-
-/**
-* @name    metaData.root
-* @description 	Name of the property containing rows ("rows"). This is mainly for the Ext grid component.
-* @type     String
-*/
-
-/**
-* @name    metaData.totalProperty
-* @description 	Name of the top-level property
-            containing the row count ("rowCount") in our case. This is mainly
-            for the Ext grid component.
-* @type   String
-*/
-
-/**
-* @name   metaData.sortInfo
-* @description  Sort specification in Ext grid terms.
-            This contains two sub-properties, field and direction, which indicate
-            the sort field and direction ("ASC" or "DESC") respectively.
-* @type   Object
-*/
-
-/**
-* @name    metaData.id
-* @description  Name of the primary key column.
-* @type     String
-*/
-
-/**
-* @name    metaData.fields
-* @description	Array of field information.
-* @type    LABKEY.Query.FieldMetaData[]
-*/
-
-/**
-* @name    columnModel
-* @description   Contains information about how one may interact
-            with the columns within a user interface. This format is generated
-            to match the requirements of the Ext grid component. See
-            <a href="http://extjs.com/deploy/ext-2.2.1/docs/?class=Ext.grid.ColumnModel">
-            Ext.grid.ColumnModel</a> for further information.
-* @type  Object[]
-*/
-
-/**
-* @name   rows
-* @description    An array of rows, each of which is a
-            sub-element/object containing a property per column.
-* @type   Object[]
-*/
-
-/**
-* @name rowCount
-* @description Indicates the number of total rows that could be
-            returned by the query, which may be more than the number of objects
-            in the rows array if the client supplied a value for the query.maxRows
-            or query.offset parameters. This value is useful for clients that wish
-            to display paging UI, such as the Ext grid.
-* @type   Integer
-*/
-
-/**#@-*/
-
- /**
-* @name LABKEY.Query.ExtendedSelectRowsResults
-* @class  ExtendedSelectRowsResults class to describe the first
-            object passed to the successCallback function by
-            {@link LABKEY.Query.selectRows} if config.requiredVersion is set to "9.1".
-  *            <p>Additional Documentation:
-  *              <ul>
-  *                  <li><a href="https://www.labkey.org/Documentation/wiki-page.view?name=findNames">
-  *                      How To Find schemaName, queryName &amp; viewName</a></li>
-  *              </ul>
-  *           </p>
-* @see LABKEY.Query.selectRows
- */
-
-/**#@+
-* @memberOf LABKEY.Query.ExtendedSelectRowsResults#
-* @field
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#metaData
-* @description Contains type and lookup information about the columns in the resultset.
-* @type    Object
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#metaData.root
-* @description 	Name of the property containing rows ("rows"). This is mainly for the Ext grid component.
-* @type     String
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#metaData.totalProperty
-* @description 	Name of the top-level property
-            containing the row count ("rowCount") in our case. This is mainly
-            for the Ext grid component.
-* @type   String
-*/
-
-/**
-* @name   LABKEY.Query.ExtendedSelectRowsResults#metaData.sortInfo
-* @description  Sort specification in Ext grid terms.
-            This contains two sub-properties, field and direction, which indicate
-            the sort field and direction ("ASC" or "DESC") respectively.
-* @type   Object
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#metaData.id
-* @description  Name of the primary key column.
-* @type     String
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#metaData.fields
-* @description	Array of field information.
-            Each field has the following properties:
-            <ul><li><b>name</b> -- The name of the field</li>
-            <li><b>type</b> -- JavaScript type name of the field</li>
-            <li><b>shownInInsertView</b> -- whether this field is intended to be shown in insert views</li>
-            <li><b>shownInUpdateView</b> -- whether this field is intended to be shown in update views</li>
-            <li><b>shownInDetailsView</b> -- whether this field is intended to be shown in details views</li>
-            <li><b>measure</b> -- whether this field is a measure.  Measures are fields that contain data subject to charting and other analysis.</li>
-            <li><b>dimension</b> -- whether this field is a dimension.  Data dimensions define logical groupings of measures.</li>
-            <li><b>hidden</b> -- whether this field is hidden and not normally shown in grid views</li>
-            <li><b>lookup</b> -- If the field is a lookup, there will
-                be four sub-properties listed under this property:
-                schema, table, displayColumn, and keyColumn, which describe the schema, table, and
-                display column, and key column of the lookup table (query).</li></ul>
-* @type    Object[]
-*/
-
-/**
-* @name    LABKEY.Query.ExtendedSelectRowsResults#columnModel
-* @description   Contains information about how one may interact
-            with the columns within a user interface. This format is generated
-            to match the requirements of the Ext grid component. See
-            <a href="http://extjs.com/deploy/ext-2.2.1/docs/?class=Ext.grid.ColumnModel">
-            Ext.grid.ColumnModel</a> for further information.
-* @type  String
-*/
-
-/**
-* @name   LABKEY.Query.ExtendedSelectRowsResults#rows
-* @description    An array of rows, each of which is a
-            sub-element/object containing an object per column. The
-            object will always contain a property named "value" that is the
-            column's value, but it may also contain other properties about
-            that column's value. For example, if the column was setup to track
-            missing value information, it will also contain a property named mvValue (which
-            is the raw value that is considered suspect), and a property named
-            mvIndicator, which will be the string MV indicator (e.g., "Q").
-* @type   Object[]
-*/
-
-/**
-* @name LABKEY.Query.ExtendedSelectRowsResults#rowCount
-* @description Indicates the number of total rows that could be
-            returned by the query, which may be more than the number of objects
-            in the rows array if the client supplied a value for the query.maxRows
-            or query.offset parameters. This value is useful for clients who wish
-            to display paging UI, such as the Ext grid.
-* @type   Integer
-*/
-
-/**#@-*/
-
-/** docs for methods defined in dom/Query.js - primarily here to ensure API docs get generated with combined core/dom versions */
-
-/**
- * Execute arbitrary LabKey SQL and export the results to Excel or TSV. After this method is
- * called, the user will be prompted to accept a file from the server, and most browsers will allow
- * the user to either save it or open it in an appropriate application.
- * For more information, see the
- * <a href="https://www.labkey.org/Documentation/wiki-page.view?name=labkeySql">
- * LabKey SQL Reference</a>.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name exportSql
- * @param config An object which contains the following configuration properties.
- * @param {String} config.schemaName name of the schema to query.
- * @param {String} config.sql The LabKey SQL to execute.
- * @param {String} [config.format] The desired export format. May be either 'excel' or 'tsv'. Defaults to 'excel'.
- * @param {String} [config.containerPath] The path to the container in which the schema and query are defined,
- *       if different than the current container. If not supplied, the current container's path will be used.
- * @param {String} [config.containerFilter] One of the values of {@link LABKEY.Query.containerFilter} that sets
- *       the scope of this query. Defaults to containerFilter.current, and is interpreted relative to
- *       config.containerPath.
- */
-
-/**
- * Load the set of user visible schemas from the given container into a standard &lt;select&gt; input element.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name schemaSelectInput
- * @param config An object which contains the following configuration properties.
- * @param {String} config.renderTo the id of the &lt;select&gt; input to load the LabKey queries into.
- * @param {String} config.initValue the initial value to try and set the &lt;select&gt; element value after it loads.
- */
-
-/**
- * Load the set of queries from this container for a given schema into a standard &lt;select&gt; input. The config object
- * must define which &lt;select&gt; input is for the schemas and which &lt;select&gt; input is for the queries. This function
- * also then associates the two &lt;select&gt; inputs so that a selection change in the schema input will update the
- * query input accordingly.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name querySelectInput
- * @param config An object which contains the following configuration properties.
- * @param {String} config.renderTo the id of the &lt;select&gt; input to load the LabKey queries into.
- * @param {String} config.schemaInputId the id of the &lt;select&gt; input to load the LabKey schemas into.
- * @param {String} config.initValue the initial value to try and set the &lt;select&gt; element value after it loads.
- */
-
-/**
- * Load the set of columns for a given schema/query into a standard &lt;select&gt; input. The config object
- * must define the schemaName and queryName to be used to source the column listing.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name columnSelectInput
- * @param config An object which contains the following configuration properties.
- * @param {String} config.renderTo the id of the &lt;select&gt; input to load the LabKey queries into. Required.
- * @param {String} config.schemaName the name of the schema. Required.
- * @param {String} config.queryName the name of the query. Required.
- * @param {String} config.initValue the initial value to try and set the &lt;select&gt; element value after it loads. Optional.
- * @param {String} config.filterFn a function to call to filter the column set (ex. by data type). Optional.
- * @param {String} config.sortFn a function to call to sort the column set. Optional.
- */
-
-/**
- * Load the set of views for a given schema/query into a standard &lt;select&gt; input. The config object
- * must define which &lt;select&gt; input is for the schemas and which &lt;select&gt; input is for the queries. This function
- * also then associates the two &lt;select&gt; inputs so that a selection change in the query input will update the
- * views input accordingly.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name queryViewSelectInput
- * @param config An object which contains the following configuration properties.
- * @param {String} config.renderTo the id of the &lt;select&gt; input to load the LabKey query views into. Required.
- * @param {String} config.schemaInputId the name of the &lt;select&gt; input to pull the LabKey schema from.
- * @param {String} config.queryInputId the name of the &lt;select&gt; input to load the LabKey query from.
- * @param {String} config.initValue the initial value to try and set the &lt;select&gt; element value after it loads. Optional.
- */
-
-/**
- * Bulk import data rows into a table.
- * One of 'text', 'path', 'moduleResource', or 'file' is required and cannot be combined.
- *
- * @memberOf LABKEY.Query
- * @function
- * @static
- * @name importData
- * @param {Object} config An object which contains the following configuration properties.
- * @param {String} config.schemaName Name of a schema defined within the current container.
- * @param {String} config.queryName Name of a query table associated with the chosen schema.
- * @param {File} [config.file] A <a href='https://developer.mozilla.org/en-US/docs/DOM/File'><code>File</code></a> object or a file input element to upload to the server.
- * @param {String} [config.text] Text to import.
- * @param {String} [config.path] Path to resource under webdav tree. E.g. "/_webdav/MyProject/@files/data.tsv"
- * @param {String} [config.module] Module name to use when resolving a module resource.
- * @param {String} [config.moduleResource] A file resource within the module to import.
- * @param {String} [config.importIdentity] When true, auto-increment key columns may be imported from the data.
- * @param {String} [config.importLookupByAlternateKey] When true, lookup columns can be imported by their alternate keys instead of the primary key.
- *          For example, if a column is a lookup to a SampleSet, the imported value can be the Sample's name since names must be unique within a SampleSet.
- * @param {String} [config.insertOption] If IMPORT, rows will be inserted and an error will be given is existing rows are present in the table.
- *          If MERGE, new rows will be inserted and existing rows will be merged.
- * @param {Function} [config.success] Function called when the "importData" function executes successfully.
- Will be called with the following arguments:
- An object containing success and rowCount properties.
- * @param {Function} [config.failure]  Function called importing data fails.
- * @param {String} [config.containerPath] The container path in which the schema and query name are defined.
- * @param {Integer} [config.timeout] The maximum number of milliseconds to allow for this operation before
- *       generating a timeout error (defaults to 30000).
- * @param {Object} [config.scope] A scope for the callback functions. Defaults to "this"
- * @returns {Mixed} In client-side scripts, this method will return a transaction id
- * for the async request that can be used to cancel the request
- * (see <a href="http://dev.sencha.com/deploy/dev/docs/?class=Ext.data.Connection&member=abort" target="_blank">Ext.data.Connection.abort</a>).
- * In server-side scripts, this method will return the JSON response object (first parameter of the success or failure callbacks.)
- * @example Example, importing tsv data from a module: <pre name="code" class="javascript">
- LABKEY.Query.importData({
-             schemaName: 'lists',
-             queryName: 'People',
-             // reference to &lt;input type='file' id='file'&gt;
-             file: document.getElementById('file')
-         },
- });</pre>
- * @example Example, importing tsv data from a module: <pre name="code" class="javascript">
- LABKEY.Query.importData({
-             schemaName: 'lists',
-             queryName: 'People',
-             module: 'mymodule',
-             moduleResource: '/data/lists/People.tsv'
-         },
- });</pre>
- */
+exports.Query = LABKEY.Query;
