--- core/script-fragments/labkey/FieldKey.header.js
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2013 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-var console = require("console");
-var LABKEY = require("./init");
