--- core/script-fragments/labkey/Filter.footer.js
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2011 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-exports.Filter = LABKEY.Filter;
