--- core/script-fragments/labkey/Message.footer.js
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2011 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-exports.Message = LABKEY.Message;
-
