--- core/script-fragments/labkey/Message.header.js
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2011-2012 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-var LABKEY = require("./init");
-LABKEY.ActionURL = require("./ActionURL").ActionURL;
-LABKEY.ExtAdapter = require("ExtAdapter").Adapter;
-
-LABKEY.Utils = require("./Utils").Utils;
