--- core/script-fragments/labkey/Query.footer.js
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2011 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-exports.Query = LABKEY.Query;
-
-
