--- core/script-fragments/labkey/Query.header.js
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2011-2012 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-var console = require("console");
-var LABKEY = require("./init");
-LABKEY.ActionURL = require("./ActionURL").ActionURL;
-LABKEY.Ajax = require("./Ajax").Ajax;
-LABKEY.ExtAdapter = require("ExtAdapter").Adapter;
-LABKEY.Filter = require("./Filter").Filter;
-LABKEY.Utils = require("./Utils").Utils;
