--- core/script-fragments/labkey/Report.footer.js
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2012-2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-exports.Report = LABKEY.Report;
-
-
