--- core/script-fragments/labkey/Report.header.js
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2012-2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-var LABKEY = require("./init");
-LABKEY.ActionURL = require("./ActionURL").ActionURL;
-LABKEY.Ajax = require("./Ajax").Ajax;
-LABKEY.Utils = require("./Utils").Utils;
