--- core/script-fragments/labkey/Security.footer.js
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2011-2012 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-LABKEY.Security.getPolicy = function () { throw Error("Not yet implemented"); }
-
-exports.Security = LABKEY.Security;
-
