--- core/script-fragments/labkey/Security.header.js
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2011-2012 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
- */
-
-var LABKEY = require("./init");
-LABKEY.ActionURL = require("./ActionURL").ActionURL;
-LABKEY.Ajax = require("./Ajax").Ajax;
-LABKEY.ExtAdapter = require("ExtAdapter").Adapter;
-LABKEY.SecurityPolicy = {}; //NYI
-LABKEY.Utils = require("./Utils").Utils;
